requirements.txt
pandas
numpy
scikit-learn
seaborn
matplotlib
pyyaml
